* Added the following LuaSyncedCtrl call-outs:     - SetUnitLosMask(unitID, allyTeamID, number | table) -> nil     - SetUnitLosState(unitID, allyTeamID, number | table) -> nil   - If a mask bit is enabled, than the engine will not update     the corresponding state bit (think of them as 'lock' bits)   - If a table is used, entries should be of the form:     { string = boolean } -- 'false' to clear, 'true' to set     ex:  SetUnitLosState(3, 2, { los = true, radar = false })   - The valid strings are 'los', 'radar', 'prevLos', and 'contRadar'       prevLos:    previous LOS       contRadar:  continuous radar  (since prevLos contact)       *** These 2 dictate whether or not a unit type is known * Replaced LOS_INTEAM with the LOS_xxx_MASK bits * Changed the losStatus array from 'int' to 'unsigned short' * Minimap and UnitDrawer now rely on the losStatus for unit   visibility, and do not take the allyTeam shortcut
* Added the 'customParams' table to both featureDefs and weaponDefs
  (accessible via the lua FeatureDefs and WeaponDefs' tables)

* Did some indent/whitespace clean-up in Game.cpp

* Globally replaced  s/Recieve/Receive


git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@5907 37977431-3df6-0310-b722-df95706aa16b

+++ rts/ExternalAI/AICallback.cpp
@@ -556,8 +556,8 @@ const UnitDef* CAICallback::GetUnitDef(int unitid)
 			if (gs->Ally(unit->allyteam, allyTeam)) {
 				return unitDef;
 			}
-			const int losStatus = unit->losStatus[allyTeam];
-			const int prevMask = (LOS_PREVLOS | LOS_CONTRADAR);
+			const unsigned short losStatus = unit->losStatus[allyTeam];
+			const unsigned short prevMask = (LOS_PREVLOS | LOS_CONTRADAR);
 			if (((losStatus & LOS_INLOS) != 0) ||
 					((losStatus & prevMask) == prevMask)) {
 				const UnitDef* decoyDef = unitDef->decoyDef;
